pyproject.toml
[project]
name = "rapid"
version = "0.1.0"
description = "Python package for RapidAPI.com"
authors = [
    {name = "Brendan A. Burgee", email = "brendanburgee@gmail.com"},
]
dependencies = ["requests>=2.32.3", "python-dotenv>=1.0.1"]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
